requirements.txt
flask==2.3.3
pdfminer.six==20221105
python-dotenv==1.0.0
SQLAlchemy==2.0.20
werkzeug==2.3.7
PyPDF2==3.0.1
